+++ src/eth/block_miner.rs
@@ -296,10 +296,7 @@ pub fn block_from_local(number: BlockNumber, txs: NonEmpty<LocalTransactionExecu
         let mut mined_logs: Vec<LogMined> = Vec::with_capacity(tx.result.execution.logs.len());
         for mined_log in tx.result.execution.logs.clone() {
             // calculate bloom
-            block.header.bloom.accrue(BloomInput::Raw(mined_log.address.as_ref()));
-            for topic in mined_log.topics().into_iter() {
-                block.header.bloom.accrue(BloomInput::Raw(topic.as_ref()));
-            }
+            block.header.bloom.accrue_log(&mined_log);
 
             // mine log
             let mined_log = LogMined {
